models/requirements.txt
torch
numpy
scipy
matplotlib
h5py
neuraloperator
torchdiffeq